Skipping git reset on deploy unless requested

+++ deploy_tools/fabfile.py
@@ -22,8 +22,9 @@ def _get_latest_source():
         run("git fetch")
     else:
         run(f"git clone {REPO_URL} .")
-    current_commit = local("git log -n 1 --format=%H", capture=True)
-    run(f"git reset --hard {current_commit}")
+    if os.getenv("CURRENT_COMMIT") == "TRUE":
+        current_commit = local("git log -n 1 --format=%H", capture=True)
+        run(f"git reset --hard {current_commit}")
 
 
 def _update_virtualenv():
